+++ CHANGELOG.md
@@ -28,8 +28,8 @@
   - Added a `[backends.<id>]` setting for configuring the backend (not the tool) itself.
 - **WASM API**
   - Added a `get_backend_config` helper function to retrieve the backend configuration.
-  - Added a `define_tool_config` plugin function.
-  - Added a `DefineToolConfigOutput` type.
+  - Added a `define_tool_config` and `define_backend_config` plugin (guest) function.
+  - Added a `DefineToolConfigOutput` and `DefineBackendConfigOutput` types.
   - Added a `ToolInventoryOptions.scoped_backend_dir` field.
 
 ## 0.52.5